+++ Driver.lua
@@ -795,12 +795,28 @@ end
 -- todo: remove this, eventually
 function TargetedSpellsDriver:EncounterPreventsTTSExecution(unit)
 	-- Lothraxion, Nexus-Point Xenas
-	if self.activeEncounterId == 3333 and UnitLevel(unit) == -1 then
-		local id = string.gsub(unit, "nameplate", "")
-		local nameplateId = tonumber(id)
-		-- ignores casts by a boss unit if the nameplateN id is greater than 1
-		-- Divine Guile spawns 12 images, all UnitLevel == -1
-		return nameplateId > 1
+	if self.activeEncounterId == 3333 then
+		if UnitLevel(unit) == -1 then
+			local id = string.gsub(unit, "nameplate", "")
+			-- ignores casts by a boss unit if the nameplateN id is greater than 1
+			-- Divine Guile spawns 12 images, all UnitLevel == -1
+			return tonumber(id) > 1
+		end
+
+		return false
+	end
+
+	-- Ick and Krick, Pit of Saron
+	if self.activeEncounterId == 2001 then
+		-- ignore Shade of Krick, Shadowbind casts
+		-- this might ignore the Gargoyle too but realistically you don't pull that (please)
+		return UnitLevel(unit) == 91 and UnitIsLieutenant(unit)
+	end
+
+	-- Viceroy Nezhar, Seat of the Triumvirate
+	if self.activeEncounterId == 2067 then
+		-- ignores Umbral Tentacle casts (and also other elites if you pull those but why would you)
+		return UnitLevel(unit) == 90
 	end
 
 	return false